requirements.txt
pandas
numpy
torch
opencv-python
Pillow
ultralytics
matplotlib
torchvision
streamlit
